Rydd opp i struktur
